Refactor creation and upgrade tests to own module
